+++ src/web/api/__init__.py
@@ -1,0 +1,3 @@
+from web.api.internal import InternalApi
+
+__all__ = ['InternalApi']